+++ st2common/tests/unit/test_db_fields.py
@@ -13,27 +13,43 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import time
 import datetime
+import calendar
 
 import mock
 import unittest2
 
 from st2common.fields import ComplexDateTimeField
+from st2common.util import isotime
 
 
 class ComplexDateTimeFieldTestCase(unittest2.TestCase):
+    def test_what_comes_in_goes_out(self):
+        field = ComplexDateTimeField()
+
+        date = datetime.datetime.utcnow()
+        date = isotime.add_utc_tz(date)
+
+        us = field._datetime_to_microseconds_since_epoch(date)
+        result = field._microseconds_since_epoch_to_datetime(us)
+        self.assertEqual(date, result)
+
     def test_round_trip_conversion(self):
         datetime_values = [
             datetime.datetime(2015, 1, 1, 15, 0, 0).replace(microsecond=500),
             datetime.datetime(2015, 1, 1, 15, 0, 0).replace(microsecond=0),
             datetime.datetime(2015, 1, 1, 15, 0, 0).replace(microsecond=999999)
         ]
+        datetime_values = [
+            isotime.add_utc_tz(datetime_values[0]),
+            isotime.add_utc_tz(datetime_values[1]),
+            isotime.add_utc_tz(datetime_values[2])
+        ]
         microsecond_values = []
 
         # Calculate microsecond values
         for value in datetime_values:
-            seconds = time.mktime(value.timetuple())
+            seconds = calendar.timegm(value.timetuple())
             microseconds_reminder = value.time().microsecond
             result = int(seconds * 1000000) + microseconds_reminder
             microsecond_values.append(result)
@@ -68,6 +84,7 @@ class ComplexDateTimeFieldTestCase(unittest2.TestCase):
 
         # Microseconds
         dt = datetime.datetime(2015, 1, 1, 15, 0, 0).replace(microsecond=500)
+        dt = isotime.add_utc_tz(dt)
         us = field._datetime_to_microseconds_since_epoch(value=dt)
         mock_get.return_value = us
         self.assertEqual(field.__get__(instance=None, owner=None), dt)